# docs/openapi.yaml
...
openapi: "3.0.3"
info:
  title: Nuqta Backend API
  version: 1.0.0
  description: |
    REST JSON API for Nuqta POS / ERP system.
    All responses use the standard envelope:
    - Success: `{ ok: true, data: T }`
    - Error:   `{ ok: false, error: { code, message, details?, status? } }`
  contact:
    name: Nuqta Dev Team
servers:
  - url: /api/v1
    description: API v1 base path

# ─── Security Schemes ──────────────────────────────────────────────────
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT HS256 token.
        Access token TTL: 15 min (900s).
        Refresh token TTL: 7 days (604800s).
        Header: `Authorization: Bearer <accessToken>`

  # ─── Reusable Schemas ─────────────────────────────────────────────
  schemas:
...
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Machine-readable error code
          enum:
            - VALIDATION_ERROR
            - UNAUTHORIZED
            - FORBIDDEN
            - NOT_FOUND
            - CONFLICT
            - IDEMPOTENCY_REPLAY
            - INSUFFICIENT_STOCK
            - INVALID_STATE
            - INTERNAL_ERROR
        message:
          type: string
        details:
          description: Additional context (validation field errors, etc.)
        status:
          type: integer
    ErrorEnvelope:
      type: object
      required: [ok, error]
      properties:
        ok:
          type: boolean
          enum: [false]
        error:
          $ref: "#/components/schemas/ApiError"

    # ── Pagination ──
...
    Product:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        sku: { type: string, nullable: true }
        barcode: { type: string, nullable: true }
        categoryId: { type: integer, nullable: true }
        description: { type: string, nullable: true }
        costPrice: { type: integer, description: "Minor units (IQD)" }
        sellingPrice: { type: integer }
        currency: { type: string }
        stock: { type: integer }
        minStock: { type: integer }
        unit: { type: string }
        supplier: { type: string, nullable: true }
        supplierId: { type: integer, nullable: true }
        expireDate: { type: string, nullable: true }
        isExpire: { type: boolean }
        status: { type: string, enum: [available, out_of_stock, discontinued] }
        isActive: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time, nullable: true }
        createdBy: { type: integer, nullable: true }
    CreateProductRequest:
      type: object
      required: [name, costPrice, sellingPrice]
      properties:
        name: { type: string, minLength: 1 }
        sku: { type: string, nullable: true }
        barcode: { type: string, nullable: true }
        categoryId: { type: integer, nullable: true }
        description: { type: string, nullable: true }
        costPrice: { type: integer, minimum: 0 }
        sellingPrice: { type: integer, minimum: 0 }
        currency: { type: string, default: IQD }
        stock: { type: integer, minimum: 0, default: 0 }
        minStock: { type: integer, minimum: 0, default: 0 }
        unit: { type: string, default: piece }
        supplier: { type: string, nullable: true }
        supplierId: { type: integer, nullable: true }
        expireDate: { type: string, format: date-time, nullable: true }
        isExpire: { type: boolean, default: false }
        status: { type: string, enum: [available, out_of_stock, discontinued], default: available }
        isActive: { type: boolean, default: true }
    UpdateProductRequest:
      type: object
      properties:
        name: { type: string }
        sku: { type: string, nullable: true }
        barcode: { type: string, nullable: true }
        categoryId: { type: integer, nullable: true }
        description: { type: string, nullable: true }
        costPrice: { type: integer, minimum: 0 }
        sellingPrice: { type: integer, minimum: 0 }
        currency: { type: string }
        stock: { type: integer, minimum: 0 }
        minStock: { type: integer, minimum: 0 }
        unit: { type: string }
        supplier: { type: string, nullable: true }
        supplierId: { type: integer, nullable: true }
        expireDate: { type: string, nullable: true }
        isExpire: { type: boolean }
        status: { type: string, enum: [available, out_of_stock, discontinued] }
        isActive: { type: boolean }
    ProductUnit:
      type: object
      properties:
        id: { type: integer }
        productId: { type: integer }
        name: { type: string }
        factor: { type: number }
        barcode: { type: string, nullable: true }
        isDefault: { type: boolean }
        createdAt: { type: string, format: date-time }
    ProductBatch:
      type: object
      properties:
        id: { type: integer }
        productId: { type: integer }
        batchNumber: { type: string }
        quantity: { type: integer }
        costPrice: { type: integer }
        expiryDate: { type: string, format: date, nullable: true }
        createdAt: { type: string, format: date-time }
    AdjustStockRequest:
      type: object
      required: [quantityChange]
      properties:
        quantityChange: { type: number }
        reason: { type: string, enum: [manual, damage, opening] }
        notes: { type: string, nullable: true }
        batchId: { type: integer, nullable: true }
        unitName: { type: string }
        unitFactor: { type: integer, minimum: 1 }
      additionalProperties: false

    # ── Customer ──
...
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema: { type: string, pattern: "^\\d+$" }
      description: Numeric resource ID
    LimitParam:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
...
    PageParam:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    SearchParam:
      name: search
      in: query
      schema: { type: string }
...
    IdempotencyKeyHeader:
      name: X-Idempotency-Key
      in: header
      required: true
      schema: { type: string, format: uuid }
      description: Idempotency key — required for financial write operations

  # ─── Reusable Responses ───────────────────────────────────────────
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    Unauthorized:
      description: Missing or invalid token
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    Conflict:
      description: Conflict (duplicate, state mismatch)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
paths:
...
  /products:
    get:
      tags: [Products]
      summary: List products (paginated)
      operationId: listProducts
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
        - $ref: "#/components/parameters/SearchParam"
        - name: categoryId
          in: query
          schema: { type: integer }
        - name: supplierId
          in: query
          schema: { type: integer }
        - name: status
          in: query
          schema: { type: string, enum: [available, out_of_stock, discontinued] }
        - name: lowStockOnly
          in: query
          schema: { type: boolean }
        - name: expiringSoonOnly
          in: query
          schema: { type: boolean }
      responses:
        "200":
          description: Paginated product list
          content:
            application/json:
              schema:
                type: object
                required: [ok, data]
                properties:
                  ok: { type: boolean, enum: [true] }
                  data:
                    type: object
                    properties:
                      items: { type: array, items: { $ref: "#/components/schemas/Product" } }
                      total: { type: integer }
                      page: { type: integer }
                      limit: { type: integer }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [Products]
      summary: Create product
      operationId: createProduct
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateProductRequest" }
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema:
                type: object
                required: [ok, data]
                properties:
                  ok: { type: boolean, enum: [true] }
                  data: { $ref: "#/components/schemas/Product" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409": { $ref: "#/components/responses/Conflict" }
  /products/{id}:
    get:
      tags: [Products]
      summary: Get product by ID
      operationId: getProduct
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Product details
          content:
            application/json:
              schema:
                type: object
                required: [ok, data]
                properties:
                  ok: { type: boolean, enum: [true] }
                  data: { $ref: "#/components/schemas/Product" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [Products]
      summary: Update product
      operationId: updateProduct
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateProductRequest" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                type: object
                required: [ok, data]
                properties:
                  ok: { type: boolean, enum: [true] }
                  data: { $ref: "#/components/schemas/Product" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [Products]
      summary: Delete product
      operationId: deleteProduct
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean, enum: [true] }
                  data: { type: "null" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /products/{productId}/units:
    get:
      tags: [Products]
      summary: List product units
      operationId: listProductUnits
      parameters:
        - name: productId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Product units
          content:
            application/json:
              schema:
                type: object
                required: [ok, data]
                properties:
                  ok: { type: boolean, enum: [true] }
                  data: { type: array, items: { $ref: "#/components/schemas/ProductUnit" } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    post:
      tags: [Products]
      summary: Create product unit
      operationId: createProductUnit
      parameters:
        - name: productId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, factor]
              properties:
                name: { type: string }
                factor: { type: number, minimum: 0.001 }
                barcode: { type: string, nullable: true }
      responses:
        "200":
          description: Created unit
          content:
            application/json:
              schema:
                type: object
                required: [ok, data]
                properties:
                  ok: { type: boolean, enum: [true] }
                  data: { $ref: "#/components/schemas/ProductUnit" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /products/units/{id}:
    put:
      tags: [Products]
      summary: Update product unit
      operationId: updateProductUnit
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                factor: { type: number }
                barcode: { type: string, nullable: true }
      responses:
        "200":
          description: Updated unit
          content:
            application/json:
              schema:
                type: object
                required: [ok, data]
                properties:
                  ok: { type: boolean, enum: [true] }
                  data: { $ref: "#/components/schemas/ProductUnit" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [Products]
      summary: Delete product unit
      operationId: deleteProductUnit
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean, enum: [true] }
                  data: { type: "null" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /products/{productId}/units/{unitId}/set-default:
    post:
      tags: [Products]
      summary: Set default unit for product
      operationId: setDefaultProductUnit
      parameters:
        - name: productId
          in: path
          required: true
          schema: { type: integer }
        - name: unitId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Default unit set
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean, enum: [true] }
                  data: { $ref: "#/components/schemas/ProductUnit" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /products/{productId}/batches:
    get:
      tags: [Products]
      summary: List product batches
      operationId: listProductBatches
      parameters:
        - name: productId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Product batches
          content:
            application/json:
              schema:
                type: object
                required: [ok, data]
                properties:
                  ok: { type: boolean, enum: [true] }
                  data: { type: array, items: { $ref: "#/components/schemas/ProductBatch" } }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [Products]
      summary: Create product batch
      operationId: createProductBatch
      parameters:
        - name: productId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [batchNumber, quantity, costPrice]
              properties:
                batchNumber: { type: string }
                quantity: { type: integer, minimum: 1 }
                costPrice: { type: integer, minimum: 0 }
                expiryDate: { type: string, format: date, nullable: true }
      responses:
        "200":
          description: Created batch
          content:
            application/json:
              schema:
                type: object
                required: [ok, data]
                properties:
                  ok: { type: boolean, enum: [true] }
                  data: { $ref: "#/components/schemas/ProductBatch" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /products/{productId}/adjust-stock:
    post:
      tags: [Products]
      summary: Adjust product stock
      operationId: adjustProductStock
      description: Used by inventoryClient for manual stock adjustments
      parameters:
        - name: productId
          in: path
          required: true
          schema: { type: integer }
        - $ref: "#/components/parameters/IdempotencyKeyHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdjustStockRequest" }
      responses:
        "200":
          description: Stock adjusted
          content:
            application/json:
              schema:
                type: object
                required: [ok, data]
                properties:
                  ok: { type: boolean, enum: [true] }
                  data: { type: object }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }

  # ─── SALES ────────────────────────────────────────────────────────
...